add oracle documentation
